Fixed expense deletion

+++ src/index.js
@@ -1152,40 +1152,43 @@ document.getElementById('selected-group').addEventListener('click',function(even
 })
 
 // expense deletion
-document.getElementById('list-expenses')?.addEventListener('click',function(event) {
-    if (event.target && event.target.classList.contains('delete')) {
-        // console.log("The expense delete btn was clicked")
-        const listItem = event.target.closest('li');
-        const expenseName = listItem.childNodes[0].childNodes[0].innerText.trim();
-        const expenseId = listItem.id;
-        // console.log(expenseName);
-        const confirmDelete = confirm(`Are you sure you want to delete: ${expenseName}`)
-        if (confirmDelete) {
-            let groupIndex = -1;
+if(listExpenses) {
+    listExpenses.addEventListener('click', function (event) {
+        if (event.target && event.target.classList.contains('delete')) {
+            // console.log("The expense delete btn was clicked")
+            const listItem = event.target.closest('li');
+            console.log(listItem);
+            // const expenseName = listItem.childNodes[0].childNodes[0].innerText.trim();
+            const expenseName = listItem.querySelector('.balances-expenses-header span').innerText;
+            const expenseId = listItem.id;
+            console.log(expenseName);
+            const confirmDelete = confirm(`Are you sure you want to delete: ${expenseName}`)
+            if (confirmDelete) {
+                let groupIndex = -1;
 
-            groupsArr.forEach((group, index) => {
-                const expenseExists = group.expenses.find(expense => expense.date == expenseId);
-                if (expenseExists) {
-                    groupIndex = index;
-                }
-            });
+                groupsArr.forEach((group, index) => {
+                    const expenseExists = group.expenses.find(expense => expense.date == expenseId);
+                    if (expenseExists) {
+                        groupIndex = index;
+                    }
+                });
 
-            if (groupIndex !== -1) {
-                const expenseIndex = groupsArr[groupIndex].expenses.findIndex(expense => expense.date == expenseId);
+                if (groupIndex !== -1) {
+                    const expenseIndex = groupsArr[groupIndex].expenses.findIndex(expense => expense.date == expenseId);
 
-                if (expenseIndex !== -1) {
-                    groupsArr[groupIndex].expenses.splice(expenseIndex, 1);
-                    localStorage.setItem('groups', JSON.stringify(groupsArr));
+                    if (expenseIndex !== -1) {
+                        groupsArr[groupIndex].expenses.splice(expenseIndex, 1);
+                        localStorage.setItem('groups', JSON.stringify(groupsArr));
 
-                    listExpenses.innerHTML = getExpensesHTML(groupsArr[groupIndex]);
+                        listExpenses.innerHTML = getExpensesHTML(groupsArr[groupIndex]);
 
+                    }
                 }
             }
+
         }
-
-    }
-});
-
+    });
+}
 selectedGroup.addEventListener('click',function(event){
     if(event.target&&event.target.id==='add-member-btn')
     {
